Restore DamageHighlight
